# docs/knowledge/backend/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/visitors/check-by-phone:
    get:
      operationId: VisitorsController_findVisitorByPhone
      parameters:
        - name: phone
          required: true
          in: query
          description: Visitor's 10-digit phone number
          schema:
            type: string
        - name: branchId
          required: true
          in: query
          description: The ID of the branch being visited
          schema:
            type: number
      responses:
        "200":
          description: ""
      summary: Check if a visitor exists by phone number for a specific branch
      tags: &a7
        - Visitor
  /api/visitors/register:
    post:
      operationId: VisitorsController_registerVisitor
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterVisitorDto"
      responses:
        "201":
          description: ""
      security:
        - access-token: []
      summary: Register a new visitor with optional documents
      tags: *a7
  /api/visitors/request-visit:
    post:
      operationId: VisitorsController_createVisitRequest
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateVisitRequestDto"
      responses:
        "201":
          description: ""
      security:
        - access-token: []
      summary: Create a visit request for a registered visitor (by Security)
      tags: *a7
...
  /api/visitors/update/{visitorId}:
    patch:
      operationId: VisitorsController_updateVisitor
      parameters:
        - name: visitorId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterVisitorDto"
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Update visitor details and documents
      tags: *a7
...
  /api/public/visitors/check-by-phone:
    get:
      operationId: PublicVisitorsController_findVisitorByPhone
      parameters:
        - name: phone
          required: true
          in: query
          description: Visitor's 10-digit phone number
          schema:
            type: string
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Check if a visitor exists by phone number for a specific branch
      tags: &a8
        - Public Visitor
  /api/public/visitors/branch-info:
    get:
      operationId: PublicVisitorsController_getBranchInfo
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Get branch information and available staff/departments
      tags: *a8
  /api/public/visitors/register:
    post:
      operationId: PublicVisitorsController_registerLegacyVisitor
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublicRegisterVisitorDto"
      responses:
        "201":
          description: ""
      summary: Register a new visitor via public form (full registration for meeting
        visitors)
      tags: *a8
  /api/public/visitors/quick-register:
    post:
      operationId: PublicVisitorsController_quickRegisterVisitor
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuickRegisterVisitorDto"
      responses:
        "201":
          description: ""
      summary: "Quick registration for delivery visitors (minimal fields: name + phone)"
      tags: *a8
  /api/public/visitors/complete-profile/{visitorId}:
    patch:
      operationId: PublicVisitorsController_completeProfile
      parameters:
        - name: visitorId
          required: true
          in: path
          schema:
            type: string
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CompleteProfileDto"
      responses:
        "200":
          description: ""
      summary: Complete visitor profile for meeting visits (upgrade from delivery
        profile)
      tags: *a8
...
  /api/public/visitors/meeting-visit:
    post:
      operationId: PublicVisitorsController_createMeetingVisitRequest
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingVisitRequestDto"
      responses:
        "201":
          description: ""
      summary: Create a meeting visit request (requires staff approval)
      tags: *a8
  /api/public/visitors/request-visit:
    post:
      operationId: PublicVisitorsController_createVisitRequest
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublicCreateVisitRequestDto"
      responses:
        "201":
          description: ""
      summary: Create a visit request via public form (legacy endpoint)
      tags: *a8
  /api/public/visitors/update/{visitorId}:
    patch:
      operationId: PublicVisitorsController_updateVisitor
      parameters:
        - name: visitorId
          required: true
          in: path
          schema:
            type: string
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublicRegisterVisitorDto"
      responses:
        "200":
          description: ""
      summary: Update visitor details via public form (from QR code)
      tags: *a8
  /api/public/visitors/send-otp:
    post:
      description: Initiates phone verification by sending a 6-digit OTP via SMS.
        Works for both existing and new visitors.
      operationId: PublicVisitorsController_sendOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpDto"
      responses:
        "201":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponseDto"
        "400":
          description: Bad request - validation failed, OTP locked, or SMS send failed
        "429":
          description: Too many requests - rate limit exceeded
      summary: Send OTP to visitor phone for verification
      tags: *a8
  /api/public/visitors/verify-phone:
    post:
      description: Validates the OTP sent to the visitor phone. Returns visitor data
        and indicates if visitor is existing or new.
      operationId: PublicVisitorsController_verifyPhone
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPhoneDto"
      responses:
        "200":
          description: OTP verified successfully
        "400":
          description: Invalid OTP, expired OTP, or locked OTP
        "404":
          description: Visitor not found
      summary: Verify visitor phone number using OTP
      tags: *a8
  /api/public/visitors:
    post:
      description: Creates or updates visitor record and creates a pending visit
        request. Requires phone verification first.
      operationId: PublicVisitorsController_registerVisitor
      parameters: []
      responses:
        "201":
          description: Visit created successfully
        "400":
          description: Validation failed or phone not verified
        "404":
          description: Visitor or branch not found
        "500":
          description: File upload or database error
      summary: Register visitor and create visit request
      tags: *a8
...
  /api/staff/pending-visits:
    get:
      operationId: StaffController_getPendingVisits
      parameters: []
      responses:
        "200":
          description: ""
      security: &a12
        - access-token: []
      summary: Get all of current staff member's pending visitor requests
      tags: &a13
        - Staff Dashboard
  /api/staff/history:
    get:
      operationId: StaffController_getVisitorHistory
      parameters: []
      responses:
        "200":
          description: ""
      security: *a12
      summary: Get all of current staff member's recent visitor history
      tags: *a13
  /api/staff/visits/{visitId}/approve:
    patch:
      operationId: StaffController_approveVisit
      parameters:
        - name: visitId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a12
      summary: Approve a visitor request
      tags: *a13
  /api/staff/visits/{visitId}/reject:
    patch:
      operationId: StaffController_rejectVisit
      parameters:
        - name: visitId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectVisitDto"
      responses:
        "200":
          description: ""
      security: *a12
      summary: Reject a visitor request
      tags: *a13
  /api/security/visits/{visitId}/approve:
    patch:
      operationId: SecurityController_approveVisit
      parameters:
        - name: visitId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: &a14
        - access-token: []
      summary: Approve a visitor request (by Security)
      tags: &a15
        - Security Dashboard
  /api/security/visits/{visitId}/reject:
    patch:
      operationId: SecurityController_rejectVisit
      parameters:
        - name: visitId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectVisitDto"
      responses:
        "200":
          description: ""
      security: *a14
      summary: Reject a visitor request (by Security)
      tags: *a15
  /api/security/visitors/counts:
    get:
      operationId: SecurityController_getVisitorCounts
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID to filter visitors
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a14
      summary: Get visitor counts by status for security dashboard
      tags: *a15
  /api/security/visitors:
    get:
      operationId: SecurityController_getVisitorsByStatus
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID to filter visitors
          schema:
            type: string
        - name: status
          required: true
          in: query
          description: Status values to filter by
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: ""
      security: *a14
      summary: Get visitors filtered by status for security dashboard
      tags: *a15
  /api/security/visits/{visitId}/details:
    get:
      description: Returns full visitor details including unmasked phone, email,
        company, designation, and all timestamps
      operationId: SecurityController_getVisitorDetails
      parameters:
        - name: visitId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Visitor details retrieved successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Visit not found
      security: *a14
      summary: Get complete visitor details for security dashboard
      tags: *a15
...
info:
  title: Hospital Visitor Tracking System
  description: API documentation for the Hospital Visitor Tracking System
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    RegisterVisitorDto:
      type: object
      properties:
        phone:
          type: string
        branchId:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        alternatePhone:
          type: string
        email:
          type: string
        alternateEmail:
          type: string
        address:
          type: string
        company:
          type: string
        companyWebsite:
          type: string
        designation:
          type: string
        photo:
          type: string
          format: binary
          description: Visitor photo (.jpg, .png)
        governmentIdDocument:
          type: string
          format: binary
          description: Government ID Document (.jpg, .png, .pdf)
        officeIdDocument:
          type: string
          format: binary
          description: Office ID Document (.jpg, .png, .pdf)
        reportingManagerName:
          type: string
          description: Reporting Manager name
        reportingManagerPhone:
          type: string
          description: Reporting Manager phone
      required:
        - phone
        - branchId
        - firstName
        - lastName
    CreateVisitRequestDto:
      type: object
      properties:
        phone:
          type: string
          description: Visitor's registered phone number
        purpose:
          type: string
        department:
          type: string
          enum:
            - GENERAL_MEDICINE
            - CARDIOLOGY
            - NEUROLOGY
            - PULMONOLOGY
            - NEPHROLOGY
            - GASTROENTEROLOGY
            - ENDOCRINOLOGY
            - ONCOLOGY
            - HEMATOLOGY
            - DERMATOLOGY
            - INFECTIOUS_DISEASE
            - RHEUMATOLOGY
            - OPHTHALMOLOGY
            - DENTISTRY
            - PAIN_MANAGEMENT
            - GENERAL_SURGERY
            - ORTHOPEDICS
            - NEUROSURGERY
            - CARDIOTHORACIC_SURGERY
            - ENT_SURGERY
            - PLASTIC_RECONSTRUCTIVE_SURGERY
            - UROLOGY
            - OBSTETRICS_GYNECOLOGY
            - PEDIATRICS
            - NEONATOLOGY
            - EMERGENCY
            - INTENSIVE_CARE_UNIT
            - CRITICAL_CARE_UNIT
            - TRAUMA_CENTER
            - RADIOLOGY
            - PATHOLOGY
            - PHARMACY
            - BLOOD_BANK
            - ANESTHESIOLOGY
            - RESPIRATORY_CARE
            - CENTRAL_SUPPLY
            - NUTRITION_AND_DIETETICS
            - MEDICAL_RECORDS
            - SECURITY
            - ADMINISTRATION
            - HUMAN_RESOURCES
            - FINANCE
            - IT_SUPPORT
            - LEGAL_COMPLIANCE
        personToMeet:
          type: string
          description: ID of the staff member to meet, or "other" for unlisted staff.
            Required if staffName and staffPhone are not provided.
        staffName:
          type: string
          description: Name of the staff to meet. Required if personToMeet is "other" or
            not provided.
        staffPhone:
          type: string
          description: Phone of the staff to meet. Required if personToMeet is "other" or
            not provided.
        visitingCardPhoto:
          type: string
          description: Base64 encoded visiting card photo
      required:
        - phone
        - purpose
...
    PublicRegisterVisitorDto:
      type: object
      properties:
        phone:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        alternatePhone:
          type: string
        email:
          type: string
        alternateEmail:
          type: string
        address:
          type: string
        company:
          type: string
        companyWebsite:
          type: string
        designation:
          type: string
        reportingManagerName:
          type: string
          description: Reporting Manager name
        reportingManagerPhone:
          type: string
          description: Reporting Manager phone
        photo:
          type: string
          format: binary
          description: Visitor photo (.jpg, .png)
        governmentIdDocument:
          type: string
          format: binary
          description: Government ID Document (.jpg, .png, .pdf)
        officeIdDocument:
          type: string
          format: binary
          description: Office ID Document (.jpg, .png, .pdf)
      required:
        - phone
        - firstName
        - lastName
    QuickRegisterVisitorDto:
      type: object
      properties:
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required:
        - phone
        - firstName
        - lastName
    CompleteProfileDto:
      type: object
      properties:
        email:
          type: string
          description: Email address (required for meeting visits)
        company:
          type: string
          description: Company name (required for meeting visits)
        designation:
          type: string
          description: Designation (required for meeting visits)
        middleName:
          type: string
        alternatePhone:
          type: string
        alternateEmail:
          type: string
        companyWebsite:
          type: string
        address:
          type: string
        reportingManagerName:
          type: string
        reportingManagerPhone:
          type: string
      required:
        - email
        - company
        - designation
...
    CreateMeetingVisitRequestDto:
      type: object
      properties:
        phone:
          type: string
          description: Visitor's registered phone number
        visitSubType:
          type: string
          description: Meeting sub-type (MEDICAL_REPRESENTATIVE, SALES_MARKETING, etc.)
        purpose:
          type: string
        department:
          type: string
          enum:
            - GENERAL_MEDICINE
            - CARDIOLOGY
            - NEUROLOGY
            - PULMONOLOGY
            - NEPHROLOGY
            - GASTROENTEROLOGY
            - ENDOCRINOLOGY
            - ONCOLOGY
            - HEMATOLOGY
            - DERMATOLOGY
            - INFECTIOUS_DISEASE
            - RHEUMATOLOGY
            - OPHTHALMOLOGY
            - DENTISTRY
            - PAIN_MANAGEMENT
            - GENERAL_SURGERY
            - ORTHOPEDICS
            - NEUROSURGERY
            - CARDIOTHORACIC_SURGERY
            - ENT_SURGERY
            - PLASTIC_RECONSTRUCTIVE_SURGERY
            - UROLOGY
            - OBSTETRICS_GYNECOLOGY
            - PEDIATRICS
            - NEONATOLOGY
            - EMERGENCY
            - INTENSIVE_CARE_UNIT
            - CRITICAL_CARE_UNIT
            - TRAUMA_CENTER
            - RADIOLOGY
            - PATHOLOGY
            - PHARMACY
            - BLOOD_BANK
            - ANESTHESIOLOGY
            - RESPIRATORY_CARE
            - CENTRAL_SUPPLY
            - NUTRITION_AND_DIETETICS
            - MEDICAL_RECORDS
            - SECURITY
            - ADMINISTRATION
            - HUMAN_RESOURCES
            - FINANCE
            - IT_SUPPORT
            - LEGAL_COMPLIANCE
        personToMeet:
          type: string
          description: ID of the staff member to meet, or "other" for unlisted staff.
        staffName:
          type: string
          description: Name of the staff to meet (if personToMeet is "other").
        staffPhone:
          type: string
          description: Phone of the staff to meet (if personToMeet is "other").
        visitingCardPhoto:
          type: string
          description: Base64 encoded visiting card photo
      required:
        - phone
        - visitSubType
        - purpose
    PublicCreateVisitRequestDto:
      type: object
      properties:
        phone:
          type: string
          description: Visitor's registered phone number
        purpose:
          type: string
        department:
          type: string
          enum:
            - GENERAL_MEDICINE
            - CARDIOLOGY
            - NEUROLOGY
            - PULMONOLOGY
            - NEPHROLOGY
            - GASTROENTEROLOGY
            - ENDOCRINOLOGY
            - ONCOLOGY
            - HEMATOLOGY
            - DERMATOLOGY
            - INFECTIOUS_DISEASE
            - RHEUMATOLOGY
            - OPHTHALMOLOGY
            - DENTISTRY
            - PAIN_MANAGEMENT
            - GENERAL_SURGERY
            - ORTHOPEDICS
            - NEUROSURGERY
            - CARDIOTHORACIC_SURGERY
            - ENT_SURGERY
            - PLASTIC_RECONSTRUCTIVE_SURGERY
            - UROLOGY
            - OBSTETRICS_GYNECOLOGY
            - PEDIATRICS
            - NEONATOLOGY
            - EMERGENCY
            - INTENSIVE_CARE_UNIT
            - CRITICAL_CARE_UNIT
            - TRAUMA_CENTER
            - RADIOLOGY
            - PATHOLOGY
            - PHARMACY
            - BLOOD_BANK
            - ANESTHESIOLOGY
            - RESPIRATORY_CARE
            - CENTRAL_SUPPLY
            - NUTRITION_AND_DIETETICS
            - MEDICAL_RECORDS
            - SECURITY
            - ADMINISTRATION
            - HUMAN_RESOURCES
            - FINANCE
            - IT_SUPPORT
            - LEGAL_COMPLIANCE
        personToMeet:
          type: string
          description: ID of the staff member to meet, or "other" for unlisted staff.
            Required if staffName and staffPhone are not provided.
        staffName:
          type: string
          description: Name of the staff to meet. Required if personToMeet is "other" or
            not provided.
        staffPhone:
          type: string
          description: Phone of the staff to meet. Required if personToMeet is "other" or
            not provided.
        visitingCardPhoto:
          type: string
          description: Base64 encoded visiting card photo
      required:
        - phone
        - purpose
    SendOtpDto:
      type: object
      properties:
        phone:
          type: string
          description: Visitor's 10-digit phone number
          example: "9999999999"
        branchId:
          type: string
          description: Branch ID from QR code (UUID v4 format)
          example: 550e8400-e29b-41d4-a716-446655440000
      required:
        - phone
        - branchId
    SendOtpResponseDto:
      type: object
      properties:
        success:
          type: boolean
          description: Success flag
          example: true
        message:
          type: string
          description: User-friendly message
          example: A 6-digit code has been sent to your mobile via SMS.
        isNewVisitor:
          type: boolean
          description: Indicates if this is a new visitor (first time at this branch)
          example: false
        testOtp:
          type: string
          description: Test OTP value (only included when TEST_MODE=true)
          example: "123456"
      required:
        - success
        - message
        - isNewVisitor
    VerifyPhoneDto:
      type: object
      properties:
        phone:
          type: string
          description: Visitor's 10-digit phone number
          example: "9876543210"
        otp:
          type: string
          description: 6-digit OTP received via SMS
          example: "123456"
        branchId:
          type: string
          description: Branch ID from QR code (UUID v4 format)
          example: 550e8400-e29b-41d4-a716-446655440000
      required:
        - phone
        - otp
        - branchId
...
    RejectVisitDto:
      type: object
      properties:
        rejectionReason:
          type: string
          description: The reason for rejecting the visitor request.
          example: Meeting has been rescheduled.
      required:
        - rejectionReason